+++ gradio/blocks.py
@@ -89,7 +89,7 @@ class Block:
             Context.block.add(self)
         if Context.root_block is not None:
             Context.root_block.blocks[self._id] = self
-            if isinstance(self, components.TempFileManager):
+            if isinstance(self, components.IOComponent):
                 Context.root_block.temp_file_sets.append(self.temp_files)
         return self
