+++ test/test_components.py
@@ -2602,3 +2602,86 @@ class TestCode:
             "interactive": None,
             "root_url": None,
         }
+
+
+class TestTempFileManagement:
+    def test_hash_file(self):
+        temp_file_manager = gr.File()
+        h1 = temp_file_manager.hash_file("gradio/test_data/cheetah1.jpg")
+        h2 = temp_file_manager.hash_file("gradio/test_data/cheetah1-copy.jpg")
+        h3 = temp_file_manager.hash_file("gradio/test_data/cheetah2.jpg")
+        assert h1 == h2
+        assert h1 != h3
+
+    @patch("shutil.copy2")
+    def test_make_temp_copy_if_needed(self, mock_copy):
+        temp_file_manager = gr.File()
+
+        f = temp_file_manager.make_temp_copy_if_needed("gradio/test_data/cheetah1.jpg")
+        try:  # Delete if already exists from before this test
+            os.remove(f)
+        except OSError:
+            pass
+
+        f = temp_file_manager.make_temp_copy_if_needed("gradio/test_data/cheetah1.jpg")
+        assert mock_copy.called
+        assert len(temp_file_manager.temp_files) == 1
+        assert Path(f).name == "cheetah1.jpg"
+
+        f = temp_file_manager.make_temp_copy_if_needed("gradio/test_data/cheetah1.jpg")
+        assert len(temp_file_manager.temp_files) == 1
+
+        f = temp_file_manager.make_temp_copy_if_needed(
+            "gradio/test_data/cheetah1-copy.jpg"
+        )
+        assert len(temp_file_manager.temp_files) == 2
+        assert Path(f).name == "cheetah1-copy.jpg"
+
+    def test_base64_to_temp_file_if_needed(self):
+        temp_file_manager = gr.File()
+
+        base64_file_1 = media_data.BASE64_IMAGE
+        base64_file_2 = media_data.BASE64_AUDIO["data"]
+
+        f = temp_file_manager.base64_to_temp_file_if_needed(base64_file_1)
+        try:  # Delete if already exists from before this test
+            os.remove(f)
+        except OSError:
+            pass
+
+        f = temp_file_manager.base64_to_temp_file_if_needed(base64_file_1)
+        assert len(temp_file_manager.temp_files) == 1
+
+        f = temp_file_manager.base64_to_temp_file_if_needed(base64_file_1)
+        assert len(temp_file_manager.temp_files) == 1
+
+        f = temp_file_manager.base64_to_temp_file_if_needed(base64_file_2)
+        assert len(temp_file_manager.temp_files) == 2
+
+        for file in temp_file_manager.temp_files:
+            os.remove(file)
+
+    @pytest.mark.flaky
+    @patch("shutil.copyfileobj")
+    def test_download_temp_copy_if_needed(self, mock_copy):
+        temp_file_manager = gr.File()
+        url1 = "https://raw.githubusercontent.com/gradio-app/gradio/main/gradio/test_data/test_image.png"
+        url2 = "https://raw.githubusercontent.com/gradio-app/gradio/main/gradio/test_data/cheetah1.jpg"
+
+        f = temp_file_manager.download_temp_copy_if_needed(url1)
+        try:  # Delete if already exists from before this test
+            os.remove(f)
+        except OSError:
+            pass
+
+        f = temp_file_manager.download_temp_copy_if_needed(url1)
+        assert mock_copy.called
+        assert len(temp_file_manager.temp_files) == 1
+
+        f = temp_file_manager.download_temp_copy_if_needed(url1)
+        assert len(temp_file_manager.temp_files) == 1
+
+        f = temp_file_manager.download_temp_copy_if_needed(url2)
+        assert len(temp_file_manager.temp_files) == 2
+
+
